// types/garjas.ts
export interface Garjas {
  id: number;
  tanggal: string;
  lari: number;
  garjas_a: number;
  pullup: number;
  shuttle_run: number;
  renang: number;
  situp: number;
  pushup: number;
  golongan: number;
  jenis_kelamin: string;
  keterangan: string;
  skor_b1: number;
  skor_b2: number;
  skor_b3: number;
  skor_b4: number;
  garjas_b: number;
  tinggi: number;
  berat: number;
  klasifikasi_tinggi_berat: string;
  skor_bmi: string;
  skor_renang: number;
  nilai_akhir: number;
  prajurit_id: string;
  created_at: string;
  updated_at: string;
}